refactor(types): split PipelineContext god-file into Types/Domain/*
Phase 1 (pivoted) — decouples the 914-LoC PipelineContext.ts god-file
into 17 focused domain modules under src/Scrapers/Pipeline/Types/Domain/.
Public API is byte-identical via the existing barrel re-exports.

* PipelineContext.ts: 914 to 270 LoC (about 70 percent reduction)
* New domain files (each <=100 LoC, single responsibility):
  - AccountDiscoveryTypes, ApiFetchContext, ApiStrategy
  - AuthDiscoveryTypes, BalanceTypes, BillingCycleTypes
  - BrowserState, DashboardState, DiagnosticsState
  - LoginState, LoginTypes, OtpTypes, PreLoginTypes
  - ScrapeDiscoveryTypes, ScrapeState
  - TxnEndpointTypes, TxnHarvestTypes
* Kept in PipelineContext.ts: IActionContext, IPipelineContext,
  IBootstrapContext (the actual sealed god-contracts)
* eslint.config.mjs: disable import-x/prefer-default-export for the
  Types/** override (type-only modules naturally have one named export)

No behaviour change. BasePhase + every phase action() untouched.
All 4810 pipeline tests still pass; tsc + eslint + architecture +
dead-code clean (282 importers, +14 from baseline 268).

Plan-pivot note: the original Phase 1 plan called for splitting
BasePhase<TPhaseContext> generically against 14 per-phase contexts.
That shape does NOT exist in the real BasePhase (Template-Method
with abstract action(IActionContext) only). This commit delivers
the SAME decoupling INTENT via a different mechanism (file split
behind the barrel) that is compatible with the real architecture.
