+++ src/Scrapers/Pipeline/Types/Domain/DashboardState.ts
@@ -1,0 +1,8 @@
+/** Dashboard phase result context. */
+interface IDashboardState {
+  readonly isReady: boolean;
+  readonly pageUrl: string;
+  readonly trafficPrimed: boolean;
+}
+
+export type { IDashboardState };
